DROOLS-2932: Implement review feedback and resolve conflicts
