+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/producers/ScenarioGridPanelProducer.java
@@ -19,11 +19,18 @@ import javax.annotation.PostConstruct;
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 
+import org.drools.workbench.screens.scenariosimulation.client.handlers.EditScenarioSimulationGridCellKeyboardOperation;
 import org.drools.workbench.screens.scenariosimulation.client.models.ScenarioGridModel;
 import org.drools.workbench.screens.scenariosimulation.client.renderers.ScenarioGridRenderer;
 import org.drools.workbench.screens.scenariosimulation.client.widgets.ScenarioGrid;
 import org.drools.workbench.screens.scenariosimulation.client.widgets.ScenarioGridLayer;
 import org.drools.workbench.screens.scenariosimulation.client.widgets.ScenarioGridPanel;
+import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.BaseGridWidgetKeyboardHandler;
+import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.KeyboardOperationMoveDown;
+import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.KeyboardOperationMoveLeft;
+import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.KeyboardOperationMoveRight;
+import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.KeyboardOperationMoveUp;
+import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.KeyboardOperationSelectTopLeftCell;
 
 /**
  * <code>@Dependent</code> <i>Producer</i> for a given {@link ScenarioGridPanel}
@@ -43,13 +50,21 @@ public class ScenarioGridPanelProducer {
                                                            scenarioGridLayer,
                                                            new ScenarioGridRenderer(false),
                                                            scenarioGridPanel);
+
+        final BaseGridWidgetKeyboardHandler handler = new BaseGridWidgetKeyboardHandler(scenarioGridLayer);
+        handler.addOperation(new EditScenarioSimulationGridCellKeyboardOperation(scenarioGridLayer),
+                             new KeyboardOperationSelectTopLeftCell(scenarioGridLayer),
+                             new KeyboardOperationMoveLeft(scenarioGridLayer),
+                             new KeyboardOperationMoveRight(scenarioGridLayer),
+                             new KeyboardOperationMoveUp(scenarioGridLayer),
+                             new KeyboardOperationMoveDown(scenarioGridLayer));
+        scenarioGridPanel.addKeyDownHandler(handler);
+
         scenarioGridLayer.addScenarioGrid(scenarioGrid);
         scenarioGridPanel.add(scenarioGridLayer);
-
     }
 
     public ScenarioGridPanel getScenarioGridPanel() {
         return scenarioGridPanel;
     }
-
 }
