+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/widgets/ScenarioGrid.java
@@ -21,20 +21,27 @@ import java.util.stream.IntStream;
 
 import com.ait.lienzo.shared.core.types.EventPropagationMode;
 import com.google.gwt.event.shared.EventBus;
+import org.drools.workbench.screens.scenariosimulation.client.events.DisableRightPanelEvent;
+import org.drools.workbench.screens.scenariosimulation.client.events.EnableRightPanelEvent;
 import org.drools.workbench.screens.scenariosimulation.client.factories.FactoryProvider;
 import org.drools.workbench.screens.scenariosimulation.client.factories.ScenarioHeaderTextBoxSingletonDOMElementFactory;
 import org.drools.workbench.screens.scenariosimulation.client.handlers.ScenarioSimulationGridWidgetMouseEventHandler;
+import org.drools.workbench.screens.scenariosimulation.client.metadata.ScenarioHeaderMetaData;
 import org.drools.workbench.screens.scenariosimulation.client.models.ScenarioGridModel;
 import org.drools.workbench.screens.scenariosimulation.client.renderers.ScenarioGridRenderer;
 import org.drools.workbench.screens.scenariosimulation.client.resources.i18n.ScenarioSimulationEditorConstants;
 import org.drools.workbench.screens.scenariosimulation.client.utils.ScenarioSimulationBuilders;
+import org.drools.workbench.screens.scenariosimulation.client.utils.ScenarioSimulationGridHeaderUtilities;
 import org.drools.workbench.screens.scenariosimulation.client.utils.ScenarioSimulationUtils;
 import org.drools.workbench.screens.scenariosimulation.model.FactIdentifier;
 import org.drools.workbench.screens.scenariosimulation.model.FactMapping;
 import org.drools.workbench.screens.scenariosimulation.model.FactMappingType;
 import org.drools.workbench.screens.scenariosimulation.model.Scenario;
 import org.drools.workbench.screens.scenariosimulation.model.Simulation;
+import org.uberfire.ext.wires.core.grids.client.model.GridColumn;
 import org.uberfire.ext.wires.core.grids.client.model.GridColumn.ColumnWidthMode;
+import org.uberfire.ext.wires.core.grids.client.model.GridData;
+import org.uberfire.ext.wires.core.grids.client.util.ColumnIndexUtilities;
 import org.uberfire.ext.wires.core.grids.client.widget.grid.NodeMouseEventHandler;
 import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.BaseGridWidget;
 import org.uberfire.ext.wires.core.grids.client.widget.grid.impl.DefaultGridWidgetCellSelectorMouseEventHandler;
@@ -225,8 +232,52 @@ public class ScenarioGrid extends BaseGridWidget {
     public boolean adjustSelection(final SelectionExtension direction, final boolean isShiftKeyDown) {
         final boolean selectionChanged = super.adjustSelection(direction, isShiftKeyDown);
 
-        scenarioGridPanel.signalRightPanelAboutSelectedHeaderCells();
+        signalRightPanelAboutSelectedHeaderCells();
 
         return selectionChanged;
+    }
+
+    public void signalRightPanelAboutSelectedHeaderCells() {
+        eventBus.fireEvent(new DisableRightPanelEvent());
+
+        if (model.getSelectedHeaderCells().size() > 0) {
+            final GridData.SelectedCell cell = model.getSelectedHeaderCells().get(0);
+
+            final int uiColumnIndex = ColumnIndexUtilities.findUiColumnIndex(model.getColumns(),
+                                                                             cell.getColumnIndex());
+            final int uiRowIndex = cell.getRowIndex();
+
+            setSelectedColumnAndHeader(uiRowIndex, uiColumnIndex);
+
+            final GridColumn column = model.getColumns().get(uiColumnIndex);
+            if (uiRowIndex > 0 && column instanceof ScenarioGridColumn) {
+                signalRightPanelHeaderCellSelected((ScenarioGridColumn) column,
+                                                   cell,
+                                                   uiColumnIndex);
+            }
+        }
+    }
+
+    private void signalRightPanelHeaderCellSelected(final ScenarioGridColumn scenarioGridColumn,
+                                                    final GridData.SelectedCell selectedHeaderCell,
+                                                    final int uiColumnIndex) {
+        final EnableRightPanelEvent enableRightPanelEvent = getEnableRightPanelEvent(scenarioGridColumn,
+                                                                                     selectedHeaderCell,
+                                                                                     uiColumnIndex);
+        eventBus.fireEvent(enableRightPanelEvent);
+    }
+
+    private EnableRightPanelEvent getEnableRightPanelEvent(final ScenarioGridColumn scenarioGridColumn,
+                                                           final GridData.SelectedCell selectedHeaderCell,
+                                                           final int uiColumnIndex) {
+        final ScenarioHeaderMetaData scenarioHeaderMetaData =
+                ScenarioSimulationGridHeaderUtilities.getColumnScenarioHeaderMetaData(scenarioGridColumn,
+                                                                                      selectedHeaderCell.getRowIndex());
+
+        return ScenarioSimulationGridHeaderUtilities.getEnableRightPanelEvent(this,
+                                                                              scenarioGridColumn,
+                                                                              scenarioHeaderMetaData,
+                                                                              uiColumnIndex,
+                                                                              scenarioHeaderMetaData.getColumnGroup());
     }
 }
